+++ example/WitTest/abi/src/runtime.rs
@@ -1,0 +1,5 @@
+use wasmedge_sdk::Caller;
+
+// trait Runtime {
+//     fn new_by_runtime(caller: &Caller, input: Vec<WasmValue>) -> (Self, Vec<WasmValue>);
+// }
