+++ example/WitTest/abi/src/wit/mod.rs
@@ -1,0 +1,4 @@
+pub(crate) mod option;
+pub(crate) mod result;
+pub(crate) mod string;
+pub(crate) mod vec;
